+++ applications/redeye-e2e/cypress.config.js
@@ -1,8 +1,7 @@
 /* eslint-disable no-undef */
 /* eslint-disable @typescript-eslint/no-var-requires */
 const { defineConfig } = require('cypress');
-
-const fs = require('fs');
+const tasks = require('./src/support/tasks');
 const path = require('path');
 
 module.exports = defineConfig({
@@ -19,41 +18,9 @@ module.exports = defineConfig({
 		configFile: path.join(__dirname, 'reporter-config.json'),
 	},
 	e2e: {
+		// eslint-disable-next-line @typescript-eslint/no-unused-vars
 		setupNodeEvents(on, config) {
-			on('task', {
-				// ===== task to use node 'fs' (filesystem) to read directory=====
-				readdir({ dirPath }) {
-					return new Promise((resolve, reject) => {
-						try {
-							const dirData = fs.readdirSync(dirPath);
-							resolve(dirData);
-						} catch (e) {
-							reject(e);
-						}
-					});
-				},
-				getPath({ dirPath }) {
-					return new Promise((resolve, reject) => {
-						try {
-							const dirData = path.resolve(dirPath);
-							resolve(dirData);
-						} catch (e) {
-							reject(e);
-						}
-					});
-				},
-				// ===== task to use node 'fs' (filesystem) to delete downloaded files from cypress/downloads
-				// deleteDownloads({ dirPath }) {
-				// 	fs.readdir(dirPath, (err, files) => {
-				// 		for (const file of files) {
-				// 			fs.unlink(path.join(dirPath, file), (err) => {
-				// 				console.log('Removed ' + file);
-				// 			});
-				// 		}
-				// 	});
-				// 	return null;
-				// },
-			});
+			tasks(on);
 		},
 		specPattern: '**/**/e2e/**/*.cy.js',
 		supportFile: path.join(__dirname, 'src', 'support', 'e2e.js'),
