+++ applications/redeye-e2e/src/support/beacon.js
@@ -1,3 +1,5 @@
+/* eslint-disable cypress/no-unnecessary-waiting */
+
 /// <reference types="cypress" />
 /*
  *
