+++ package.json
@@ -108,7 +108,7 @@
 		"@mikro-orm/entity-generator": "5.6.6",
 		"@mikro-orm/migrations": "5.6.6",
 		"@mikro-orm/seeder": "5.6.6",
-		"@moonrepo/cli": "^0.25.1",
+		"@moonrepo/cli": "^0.25.3",
 		"@swc/cli": "~0.1.55",
 		"@swc/core": "~1.2.143",
 		"@swc/helpers": "~0.3.3",
